:sparkles: feat: DTO, Entity 재정의
Ref: #S14P11A704-190, S14P11A704-197

+++ backend/src/main/java/com/mento/domain/payment/dto/KakaoApproveResDto.java
@@ -25,15 +25,15 @@ public record KakaoApproveResDto(
 
 	@JsonProperty("approved_at")
 	LocalDateTime approvedAt
-	
+
 ) {
-	public Integer totalAmount() {
+	public Long totalAmount() {
 		return (amount == null) ? 0 : amount.total();
 	}
 
 	private record Amount(
 		@JsonProperty("total")
-		Integer total
+		Long total
 	) {
 	}
